Refactor into resolver and iterator
This is a work in progress. The code committed here works, but it's probably
not ideal for actually implementing a menu; the intended direction of this
module is a JSON dump of the site tree feeding into a React front end, and
the PHP side of this is a bit round about / too generic at the moment (maybe).