Add nested menu tree helper for mobile menu, rename to menukit
This commit completes the new implementation to replace the much-copied
old menu generation snippet.

The full backstory, for anyone interested:

A couple of years ago I wrote two functions in the Z Energy site to
facilitate rendering the entire site tree without recursing through
SiteTree::Children(), since doing so results in way too many queries.

The menu for mobile devices that this PHP code supported was quite
popular internally, and it started appearing in the code of other
projects. Unfortunately, at no point until now was the time taken to
turn the code into a reusable module, so the original snippet is sprayed
around a lot of projects - bugs and all included.

This module replaces the function of that original code, and is intended
to be used in place of it, with a new implementation of the mobile menu
to go along with it.

The original implementation is the initial commit in this repo.